+++ genericdao.go
@@ -29,8 +29,8 @@ type IRowMapper interface {
 /*----------------------------------------------------------------------*/
 
 var (
-	// GdaoErrorDuplicatedEntry indicates that the write operation failed because of data integrity violation: entry/key duplicated.
-	GdaoErrorDuplicatedEntry = errors.New("data integrity violation: duplicated entry/key")
+	// ErrGdaoDuplicatedEntry indicates that the write operation failed because of data integrity violation: entry/key duplicated.
+	ErrGdaoDuplicatedEntry = errors.New("data integrity violation: duplicated entry/key")
 )
 
 // IGenericDao defines API interface of a generic data-access-object.
@@ -62,19 +62,19 @@ type IGenericDao interface {
 
 	// GdaoCreate persists one BO to database store and returns the number of saved items.
 	//
-	// If the BO already existed, this function does not modify the existing one and should return (0, GdaoErrorDuplicatedEntry).
+	// If the BO already existed, this function does not modify the existing one and should return (0, ErrGdaoDuplicatedEntry).
 	GdaoCreate(storageId string, bo IGenericBo) (int, error)
 
 	// GdaoUpdate updates one existing BO and returns the number of updated items.
 	//
 	// If the BO does not exist, this function does not create new BO and should return (0, nil).
-	// If update causes data integrity violation, this function should return (0, GdaoErrorDuplicatedEntry).
+	// If update causes data integrity violation, this function should return (0, ErrGdaoDuplicatedEntry).
 	GdaoUpdate(storageId string, bo IGenericBo) (int, error)
 
 	// GdaoSave persists one BO to database store and returns the number of saved items.
 	//
 	// If the BO already existed, this function replaces the existing one; otherwise new BO is created.
-	// If data integrity violation occurs, this function should return (0, GdaoErrorDuplicatedEntry)
+	// If data integrity violation occurs, this function should return (0, ErrGdaoDuplicatedEntry)
 	GdaoSave(storageId string, bo IGenericBo) (int, error)
 
 	// GetRowMapper returns the IRowMapper associated with the DAO.